// frontend/src/api/models/handlerStockTakingResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { HandlerStockTakingItemResponse } from './handlerStockTakingItemResponse'

/**
 * Stock taking document with full details
 */
export interface HandlerStockTakingResponse {
  approval_note?: string
  approved_at?: string
  approved_by_id?: string
  approved_by_name?: string
  completed_at?: string
  counted_items?: number
  created_at?: string
  created_by_id?: string
  created_by_name?: string
  difference_items?: number
  id?: string
  items?: HandlerStockTakingItemResponse[]
  progress?: number
  remark?: string
  started_at?: string
  status?: string
  taking_date?: string
  taking_number?: string
  tenant_id?: string
  total_difference?: number
  total_items?: number
  updated_at?: string
  version?: number
  warehouse_id?: string
  warehouse_name?: string
}