// frontend/src/api/models/handlerStockTakingItemResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

/**
 * Stock taking item with count and difference information
 */
export interface HandlerStockTakingItemResponse {
  actual_quantity?: number
  counted?: boolean
  created_at?: string
  difference_amount?: number
  difference_qty?: number
  id?: string
  product_code?: string
  product_id?: string
  product_name?: string
  remark?: string
  stock_taking_id?: string
  system_quantity?: number
  unit?: string
  unit_cost?: number
  updated_at?: string
}